fix: 맵 hooks 및 utils 분리
