+++ src/domain/Map/Map.module.scss
@@ -6,7 +6,6 @@
   display: grid;
   grid-template-columns: 340px 1fr;
   width: 100%;
-  height: calc(100vh - 200px);
   gap: 20px;
   min-height: 0;
