+++ src/domain/Roulette/components/ResultModal/ResultModal.tsx
@@ -1,54 +1,28 @@
 'use client';
 
-import { useEffect } from 'react';
 import Image from 'next/image';
+import Button from '@/shared/components/Button';
 
 import { ResultModalProps } from './type';
-import Button from '@/shared/components/Button';
+import { useEscClose } from '../../hooks/useEscClose';
+import { useShare } from '../../hooks/useShare';
 
 import styles from './ResultModal.module.scss';
 
 const DEFAULT_IMAGE = '/foods/noimg.png';
 
 export default function ResultModal({ menu, onClose }: ResultModalProps) {
-  useEffect(() => {
-    const handleEsc = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') {
-        onClose();
-      }
-    };
+  useEscClose(onClose);
+  const share = useShare();
 
-    window.addEventListener('keydown', handleEsc);
-    return () => {
-      window.removeEventListener('keydown', handleEsc);
-    };
-  }, [onClose]);
-
-  const share = async () => {
-    if (!navigator.share) {
-      alert('공유 기능을 지원하지 않는 브라우저입니다.');
-      return;
-    }
-
-    try {
-      await navigator.share({
-        title: '오늘의 메뉴',
-        text: `오늘의 메뉴는 ${menu}입니다!`,
-      });
-    } catch (error) {
-      if (error instanceof DOMException && error.name === 'AbortError') {
-        return;
-      }
-
-      // 그 외 오류는 디버깅을 위해 출력
-      console.error('Unexpected share error:', error);
-    }
+  const handleShare = () => {
+    share('오늘의 메뉴', `오늘의 메뉴는 ${menu}입니다!`);
   };
 
   return (
-    <div className={styles['overlay']} onClick={onClose}>
+    <div className={styles.overlay} onClick={onClose}>
       <div
-        className={styles['modal']}
+        className={styles.modal}
         onClick={e => e.stopPropagation()}
         role="dialog"
         aria-modal="true"
@@ -58,13 +32,13 @@ export default function ResultModal({ menu, onClose }: ResultModalProps) {
           alt={menu}
           width={300}
           height={200}
-          className={styles['modal__image']}
+          className={styles.modal__image}
         />
 
-        <p className={styles['modal__name']}>{menu}</p>
+        <p className={styles.modal__name}>{menu}</p>
 
-        <div className={styles['modal__buttons']}>
-          <Button variant="primary" size="md" onClick={share}>
+        <div className={styles.modal__buttons}>
+          <Button variant="primary" size="md" onClick={handleShare}>
             공유하기
           </Button>
