+++ src/domain/Roulette/components/RouletteFilter/RouletteFilter.tsx
@@ -1,72 +1,22 @@
 'use client';
 
-import { useState, useEffect, useRef } from 'react';
-
 import Button from '@/shared/components/Button';
 
-import { pickMenus } from '@/shared/utils/recommend/pickMenus';
-
 import { RouletteFilterProps } from './type';
-import { FoodTypeFilter, SituationFilter } from '@/types/filters';
 
 import { FOOD_TYPE_FILTERS, SITUATION_FILTERS } from '@/shared/constants/filters';
 import { FOOD_TYPE_ICONS, SITUATION_ICONS } from '@/shared/constants/icons';
 
+import { useRouletteFilter } from '../../hooks/useRouletteFilter';
+
 import styles from './RouletteFilter.module.scss';
 
-const ALL: FoodTypeFilter = '전체';
-
 export default function RouletteFilter({ onChange, disabled = false }: RouletteFilterProps) {
-  const [mode, setMode] = useState<'food' | 'situation'>('food');
-  const [selectedFoodTypes, setSelectedFoodTypes] = useState<FoodTypeFilter[]>([ALL]);
-  const [selectedSituation, setSelectedSituation] = useState<SituationFilter | null>(null);
-  const isFirstRun = useRef(true);
-
-  const toggleFoodType = (type: FoodTypeFilter) => {
-    if (disabled) return;
-
-    const isAll = type === ALL;
-    const hasAll = selectedFoodTypes.includes(ALL);
-    const isSelected = selectedFoodTypes.includes(type);
-
-    if (isAll) return setSelectedFoodTypes([ALL]);
-    if (hasAll) return setSelectedFoodTypes([type]);
-
-    if (isSelected) {
-      const updated = selectedFoodTypes.filter(t => t !== type);
-      return setSelectedFoodTypes(updated.length ? updated : [ALL]);
-    }
-
-    setSelectedFoodTypes(prev => [...prev, type]);
-  };
-
-  const toggleSituation = (sit: SituationFilter) => {
-    if (disabled) return;
-    setSelectedSituation(prev => (prev === sit ? null : sit));
-  };
-
-  useEffect(() => {
-    if (mode === 'food') {
-      setSelectedSituation(null);
-    } else {
-      setSelectedFoodTypes([ALL]);
-    }
-  }, [mode]);
-
-  useEffect(() => {
-    if (isFirstRun.current) {
-      isFirstRun.current = false;
-      return;
-    }
-
-    const menus =
-      mode === 'food' ? pickMenus(selectedFoodTypes, null) : pickMenus([ALL], selectedSituation);
-
-    onChange(menus.map(m => m.name));
-  }, [selectedFoodTypes, selectedSituation, mode, onChange]);
+  const { mode, setMode, selectedFoodTypes, toggleFoodType, selectedSituation, toggleSituation } =
+    useRouletteFilter(onChange);
 
   return (
-    <div className={styles['filter']}>
+    <div className={styles.filter}>
       <div className={styles['filter__mode']}>
         <button
           type="button"
@@ -93,7 +43,6 @@ export default function RouletteFilter({ onChange, disabled = false }: RouletteF
         {mode === 'food' &&
           FOOD_TYPE_FILTERS.map(type => {
             const isActive = selectedFoodTypes.includes(type);
-
             return (
               <Button
                 key={type}
@@ -102,10 +51,10 @@ export default function RouletteFilter({ onChange, disabled = false }: RouletteF
                 mode={isActive ? 'fill' : 'outline'}
                 padding="8px 18px"
                 fontSize="14px"
-                className={styles['pill']}
+                className={styles.pill}
                 onClick={() => toggleFoodType(type)}
               >
-                <span className={styles['icon']}>{FOOD_TYPE_ICONS[type]}</span>
+                <span className={styles.icon}>{FOOD_TYPE_ICONS[type]}</span>
                 {type}
               </Button>
             );
@@ -114,7 +63,6 @@ export default function RouletteFilter({ onChange, disabled = false }: RouletteF
         {mode === 'situation' &&
           SITUATION_FILTERS.map(sit => {
             const isActive = selectedSituation === sit;
-
             return (
               <Button
                 key={sit}
@@ -123,10 +71,10 @@ export default function RouletteFilter({ onChange, disabled = false }: RouletteF
                 mode={isActive ? 'fill' : 'outline'}
                 padding="8px 18px"
                 fontSize="14px"
-                className={styles['pill']}
+                className={styles.pill}
                 onClick={() => toggleSituation(sit)}
               >
-                <span className={styles['icon']}>{SITUATION_ICONS[sit]}</span>
+                <span className={styles.icon}>{SITUATION_ICONS[sit]}</span>
                 {sit}
               </Button>
             );
