+++ src/domain/Roulette/components/RouletteUi/type.ts
@@ -4,3 +4,12 @@ export interface RouletteUiProps {
   onStart?: () => void;
   size?: number;
 }
+export interface UseRouletteSpinProps {
+  items: string[];
+  angle: number;
+  setAngle: React.Dispatch<React.SetStateAction<number>>;
+  spinning: boolean;
+  setSpinning: React.Dispatch<React.SetStateAction<boolean>>;
+  onResult?: (v: string) => void;
+  onStart?: () => void;
+}
