+++ src/domain/Roulette/hooks/useEscClose.ts
@@ -1,0 +1,12 @@
+import { useEffect } from 'react';
+
+export function useEscClose(onClose: () => void) {
+  useEffect(() => {
+    const handleEsc = (e: KeyboardEvent) => {
+      if (e.key === 'Escape') onClose();
+    };
+
+    window.addEventListener('keydown', handleEsc);
+    return () => window.removeEventListener('keydown', handleEsc);
+  }, [onClose]);
+}
